+++ src/libs/libcommon/quest.c
@@ -68,7 +68,7 @@ s3parse_quest(pset_t *pset, uint32 n_pset, quest_t *q, char *in_str)
     s = in_str;
     
     /* skip leading whitespace */
-    for (; *s != '\0' && isspace(*s); s++);
+    for (; *s != '\0' && isspace((int)*s); s++);
 
     if (*s == '\0')	/* Nothing to parse */
 	return s;
@@ -113,7 +113,7 @@ s3parse_quest(pset_t *pset, uint32 n_pset, quest_t *q, char *in_str)
     *sp = ' ';	/* undo set to null */
 
     /* skip whitespace */
-    for (; *s != '\0' && isspace(*s); s++);
+    for (; *s != '\0' && isspace((int)*s); s++);
 
     if (s[0] == '-') {
 	if (s[1] == '1') {
@@ -131,7 +131,7 @@ s3parse_quest(pset_t *pset, uint32 n_pset, quest_t *q, char *in_str)
     }
 
     /* skip trailing whitespace, if any */
-    for (; *s != '\0' && isspace(*s); s++);
+    for (; *s != '\0' && isspace((int)*s); s++);
 
     return s;
 }
@@ -150,7 +150,7 @@ count_quest_in_conj(pset_t *pset,
 
     t = in_str;
 
-    for (; *t != '\0' && isspace(*t); t++);
+    for (; *t != '\0' && isspace((int)*t); t++);
     if (*t == ')') {
 	E_ERROR("Empty conjunction\n");
 	
@@ -160,7 +160,7 @@ count_quest_in_conj(pset_t *pset,
 	t = s3parse_quest(pset, n_pset, q, t);
 	++n_quest;
 
-	for (; t && *t != '\0' && isspace(*t); t++);
+	for (; t && *t != '\0' && isspace((int)*t); t++);
     }
     if (t == NULL) {
 	E_ERROR("Error while parsing conjunction: %s\n", in_str);
@@ -193,7 +193,7 @@ s3parse_conj(pset_t *pset,
     if (*s == '\0') return s;
 
     /* skip leading whitespace */
-    for (; *s != '\0' && isspace(*s); s++);
+    for (; *s != '\0' && isspace((int)*s); s++);
     
     if (*s == '\0') return s;
 
@@ -206,7 +206,7 @@ s3parse_conj(pset_t *pset,
 	return NULL;
     }
 
-    for (; *s != '\0' && isspace(*s); s++);
+    for (; *s != '\0' && isspace((int)*s); s++);
 
     if (*s == '\0') {
 	E_ERROR("No terms and close paren in conjunction\n", in_str);
@@ -222,7 +222,7 @@ s3parse_conj(pset_t *pset,
 
     for (i = 0; i < n_quest; i++) {
 	s = s3parse_quest(pset, n_pset, &termlst[i], s);
-	for (; *s != '\0' && isspace(*s); s++);
+	for (; *s != '\0' && isspace((int)*s); s++);
     }
 
     assert(*s == ')');
@@ -241,7 +241,7 @@ s3cnt_q_term(char *in_str)
     s = in_str;
 
     /* skip any leading whitespace */
-    for (; *s != '\0' && isspace(*s); s++);
+    for (; *s != '\0' && isspace((int)*s); s++);
     
     /* assume everything is well-formed for the moment.
      * later processing will catch syntax errors
@@ -264,7 +264,7 @@ s3parse_comp_quest(pset_t *pset,
 
     s = in_str;
 
-    for (; *s != '\0' && isspace(*s); s++);
+    for (; *s != '\0' && isspace((int)*s); s++);
 
     if (*s == '\0') {
 	E_ERROR("Empty string seen for composite question\n");
@@ -318,7 +318,7 @@ parse_simple_q(quest_t *q,
     len = strlen(q_str);
 
     /* skip leading whitespace */
-    for (i = 0; i < len && isspace(q_str[i]); i++);
+    for (i = 0; i < len && isspace((int)q_str[i]); i++);
 
     if (i == len)
 	return;
@@ -661,9 +661,12 @@ simplify_comp_quest(comp_quest_t *q,
  * Log record.  Maintained by RCS.
  *
  * $Log$
- * Revision 1.3  2001/04/05  20:02:30  awb
- * *** empty log message ***
+ * Revision 1.4  2003/11/18  21:07:25  egouvea
+ * Got rid of warning casting the argument to "isspace".
  * 
+ * Revision 1.3  2001/04/05 20:02:30  awb
+ * *** empty log message ***
+ *
  * Revision 1.2  2000/09/29 22:35:13  awb
  * *** empty log message ***
  *
